Fix a crash after the last interface is deleted
Steps to reproduce:
1. Add a router
2. Add an internal  network and a subnet
3. Attach that subnet to the router
4. Stop cfg_agent (q-asr)
5. Detach that subnet from the router
6. Restart cfg_agent (q-asr)
7. You should see a crash

The root cause is that when the last interface is deleted, neutron server
will not send any interface dictionary to the cfg_agent but it is expected
the interface dictionary is embedded inside the router info.

This patchset is to add a check in the cfg_agent such that if the interface
dictionary is not in the router, don't process any interface check and mark
all the running interface as invalid and proceed to delete

Also, this patch fixed the unit test such that it chance ciscoconfparse and
ncclient modules to a MagicMock and this cause the rest of the unit tests
can't use the actual module functionalities.

Fixes Jira: OPENSTACK-1680

Change-Id: I41781e60e739610e95263c66b33be78c33038fab