User can only add one subnet per network in ASR1K model
Currently ASR1K plugin doesn't support multiple subnets per network.
We need to add validation check to disallow user to do that.

This patchset introduces a metacloud option "subnet_per_network" to allow user
to create a certain amount of subnet under a network.

By default, the value is -1 which means we don't check the subnet count.
For ASR1K model, we will set it as 1.

Steps to reproduce
1. Create one network (external or internal)
2. Create one subnet under that network
3. Create another subnet

Fixes Jira: OPENSTACK-1717

Change-Id: I975a9e3af7016f536e5b4ab45caa596ccc7b8b2c