Fix neutron crash when updating subnet
Steps to reproduce:
1. Create a network and a subnet
2. Update that subnet (neutron subnet-update <SUB_ID> --name <New Name>

The root cause is due to the introduce of the subnet count validation.
The same function call is used when subnet is created and updated.
For the update subnet call, caller doesn't put network_id in the subnet
dictionary but our validation function is expecting that.

The fix is to relax the checking a little bit; If there is not network_id,
we just don't do any validation and log it as warning.

Fixes Jira: 1759

Change-Id: I310c373c629ebf41b490667df8f1aff3163246fb