Instance cannot talk to metadata server after DHCP agent is restart
This bug only happens on ASR1K model.

The root cause is that when DHCP agent starts up, it will try to determine if
the subnet is an 'isolated' subnet or not. To determine if it is
isolated, Neutron will look up the subnet associated ports fixed IPs.
If one of the IPs is the gateway IP, Neutron will treat the subnet as a 'non isolated' subnet.

If it is a non isolated subnet, Neutron will not put the metadata
route (169.254.169.254) in the dnsmasq option file.

Therefore, when an instance is booted up, the route is not injected
into the instance and it cannot talk to the metadata server.

The root cause is that when user attaches the subnet to a router,
ASR1K plugin will create a port which has the same IP address as the gateway IP.
This makes DHCP agent thinks that it is not an isolated subnet.

The purpose fix for Metapod version one is that we will introduce a
metacloud option to force injecting the route to the dnsmasq option
file even though it is not an isolated subnet.

Steps to reproduce:

1. Create a newtork and a subnet.
2. Attach to a router
3. Boot an instance
4. Restart dhcp agent (cm resource neutron-dhcp-agent)
5. Boot another instance

You will see the instance from step 5 is getting stuck on requesting the metadata

Fixes Jira: OPENSTACK-1761

Change-Id: I5fed4e4f47032bd71d6d2f9a3bab46ca2a0d72c5