Fix exception from add/remove interface
Suppose user accidentally remove and add the same interface in the same sync period,
cfg agent will add the interface and then remove the interface. So when user try to
update the same interface again in the next sync period, cfg agent will throw exception
as the neutron config and router config are out of sync.

This patchset is to reorder the update sequence such that cfg agent should remove the old
port first and then add the new port.

Fixes Jira: LUNAR-13

Change-Id: Ifca19ca649e144e179aea9abdc0136a79dfdc892